#790 Move identifiers IT to identifiers-inegration-tests
