+++ gbif/identifiers/identifiers-integration-tests/src/main/java/org/gbif/pipelines/Dummy.java
@@ -1,0 +1,3 @@
+package main.java.org.gbif.pipelines;
+
+public interface Dummy {}
