+++ gbif/identifiers/identifiers-integration-tests/src/test/java/org/gbif/pipelines/ExternalResourceSuite.java
@@ -1,6 +1,5 @@
-package org.gbif.pipelines.diagnostics;
+package org.gbif.pipelines;
 
-import org.gbif.pipelines.diagnostics.resources.HbaseServer;
 import org.gbif.pipelines.diagnostics.strategy.BothStrategyIT;
 import org.gbif.pipelines.diagnostics.strategy.MaxStrategyIT;
 import org.gbif.pipelines.diagnostics.strategy.MinStrategyIT;
@@ -8,6 +7,11 @@ import org.gbif.pipelines.diagnostics.strategy.OccurrenceIdStrategyIT;
 import org.gbif.pipelines.diagnostics.strategy.TripletStrategyIT;
 import org.gbif.pipelines.diagnostics.tools.IdentifiersMigratorToolIT;
 import org.gbif.pipelines.diagnostics.tools.RepairGbifIDLookupToolIT;
+import org.gbif.pipelines.keygen.HBaseLockingKeyServiceIT;
+import org.gbif.pipelines.keygen.HbaseKeyMigratorIT;
+import org.gbif.pipelines.keygen.KeygenIT;
+import org.gbif.pipelines.keygen.common.HbaseConnectionFactoryIT;
+import org.gbif.pipelines.resources.HbaseServer;
 import org.junit.ClassRule;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
@@ -21,7 +25,11 @@ import org.junit.runners.Suite.SuiteClasses;
   OccurrenceIdStrategyIT.class,
   TripletStrategyIT.class,
   IdentifiersMigratorToolIT.class,
-  RepairGbifIDLookupToolIT.class
+  RepairGbifIDLookupToolIT.class,
+  HbaseKeyMigratorIT.class,
+  HBaseLockingKeyServiceIT.class,
+  KeygenIT.class,
+  HbaseConnectionFactoryIT.class
 })
 public class ExternalResourceSuite {
   @ClassRule public static final HbaseServer HBASE_SERVER = HbaseServer.getInstance();
