+++ gbif/identifiers/identifiers-integration-tests/src/test/java/org/gbif/pipelines/diagnostics/strategy/BothStrategyIT.java
@@ -2,16 +2,17 @@ package org.gbif.pipelines.diagnostics.strategy;
 
 import java.io.IOException;
 import java.util.Map;
-import org.gbif.pipelines.diagnostics.resources.HbaseServer;
-import org.gbif.pipelines.diagnostics.resources.HbaseStore;
 import org.gbif.pipelines.keygen.HBaseLockingKeyService;
 import org.gbif.pipelines.keygen.identifier.OccurrenceKeyBuilder;
+import org.gbif.pipelines.resources.HbaseServer;
+import org.gbif.pipelines.resources.HbaseStore;
+import org.gbif.pipelines.resources.HbaseStore.KV;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
 
-public class MinStrategyIT {
+public class BothStrategyIT {
 
   /** {@link ClassRule} requires this field to be public. */
   @ClassRule public static final HbaseServer HBASE_SERVER = HbaseServer.getInstance();
@@ -22,7 +23,7 @@ public class MinStrategyIT {
   }
 
   @Test
-  public void minTest() {
+  public void bothTest() {
 
     // State
     String datasetKey = "508089ca-ddb4-4112-b2cb-cb1bff8f39ad";
@@ -30,111 +31,97 @@ public class MinStrategyIT {
     String occId = "5760d633-2efa-4359-bbbb-635f7c200803";
     String triplet = OccurrenceKeyBuilder.buildKey("ic", "cc", "cn").orElse(null);
     long gbifId = 1L;
-    long gbifId2 = 2L;
 
     String lookupOccId = datasetKey + "|" + occId;
     String lookupTriplet = datasetKey + "|" + triplet;
 
+    String occId2 = "5760d633-2efa-4359-bbbb-635f7c200804";
+    long gbifId2 = 2L;
+
     HbaseStore.putRecords(
         HBASE_SERVER.getLookupTableStore(),
-        HbaseStore.KV.create(lookupOccId, gbifId),
-        HbaseStore.KV.create(lookupTriplet, gbifId2));
+        KV.create(lookupOccId, gbifId),
+        KV.create(lookupTriplet, gbifId),
+        KV.create(occId2, gbifId2));
 
     HBaseLockingKeyService keygenService =
         new HBaseLockingKeyService(HbaseServer.CFG, HBASE_SERVER.getConnection(), datasetKey);
 
     // When
     Map<String, Long> keysToDelete =
-        new MinStrategy().getKeysToDelete(keygenService, false, triplet, occId);
+        new BothStrategy().getKeysToDelete(keygenService, false, triplet, occId);
 
     // Should
-    Assert.assertEquals(1, keysToDelete.size());
+    Assert.assertEquals(2, keysToDelete.size());
     Assert.assertEquals(Long.valueOf(gbifId), keysToDelete.get(occId));
-  }
-
-  @Test
-  public void minEqualTest() {
-
-    // State
-    String datasetKey = "508089ca-ddb4-4112-b2cb-cb1bff8f39ad";
-
-    String occId = "5760d633-2efa-4359-bbbb-635f7c200803";
-    String triplet = OccurrenceKeyBuilder.buildKey("ic", "cc", "cn").orElse(null);
-    long gbifId = 1L;
-
-    String lookupOccId = datasetKey + "|" + occId;
-    String lookupTriplet = datasetKey + "|" + triplet;
-
-    HbaseStore.putRecords(
-        HBASE_SERVER.getLookupTableStore(),
-        HbaseStore.KV.create(lookupOccId, gbifId),
-        HbaseStore.KV.create(lookupTriplet, gbifId));
-
-    HBaseLockingKeyService keygenService =
-        new HBaseLockingKeyService(HbaseServer.CFG, HBASE_SERVER.getConnection(), datasetKey);
-
-    // When
-    Map<String, Long> keysToDelete =
-        new MinStrategy().getKeysToDelete(keygenService, false, triplet, occId);
-
-    // Should
-    Assert.assertEquals(0, keysToDelete.size());
-  }
-
-  @Test
-  public void minInvertedTest() {
-
-    // State
-    String datasetKey = "508089ca-ddb4-4112-b2cb-cb1bff8f39ad";
-
-    String occId = "5760d633-2efa-4359-bbbb-635f7c200803";
-    String triplet = OccurrenceKeyBuilder.buildKey("ic", "cc", "cn").orElse(null);
-    long gbifId = 1L;
-    long gbifId2 = 2L;
-
-    String lookupOccId = datasetKey + "|" + occId;
-    String lookupTriplet = datasetKey + "|" + triplet;
-
-    HbaseStore.putRecords(
-        HBASE_SERVER.getLookupTableStore(),
-        HbaseStore.KV.create(lookupOccId, gbifId2),
-        HbaseStore.KV.create(lookupTriplet, gbifId));
-
-    HBaseLockingKeyService keygenService =
-        new HBaseLockingKeyService(HbaseServer.CFG, HBASE_SERVER.getConnection(), datasetKey);
-
-    // When
-    Map<String, Long> keysToDelete =
-        new MinStrategy().getKeysToDelete(keygenService, false, triplet, occId);
-
-    // Should
-    Assert.assertEquals(1, keysToDelete.size());
     Assert.assertEquals(Long.valueOf(gbifId), keysToDelete.get(triplet));
   }
 
   @Test
-  public void minEmptyTest() {
+  public void bothNoCollisionTest() {
 
     // State
     String datasetKey = "508089ca-ddb4-4112-b2cb-cb1bff8f39ad";
 
     String occId = "5760d633-2efa-4359-bbbb-635f7c200803";
     String triplet = OccurrenceKeyBuilder.buildKey("ic", "cc", "cn").orElse(null);
-    long gbifId2 = 2L;
+    long gbifId = 1L;
 
     String lookupOccId = datasetKey + "|" + occId;
+    String lookupTriplet = datasetKey + "|" + triplet;
+
+    String occId2 = "5760d633-2efa-4359-bbbb-635f7c200804";
+    long gbifId2 = 2L;
 
     HbaseStore.putRecords(
-        HBASE_SERVER.getLookupTableStore(), HbaseStore.KV.create(lookupOccId, gbifId2));
+        HBASE_SERVER.getLookupTableStore(),
+        KV.create(lookupOccId, gbifId),
+        KV.create(lookupTriplet, gbifId),
+        KV.create(occId2, gbifId2));
 
     HBaseLockingKeyService keygenService =
         new HBaseLockingKeyService(HbaseServer.CFG, HBASE_SERVER.getConnection(), datasetKey);
 
     // When
     Map<String, Long> keysToDelete =
-        new MinStrategy().getKeysToDelete(keygenService, false, triplet, occId);
+        new BothStrategy().getKeysToDelete(keygenService, true, triplet, occId);
 
     // Should
     Assert.assertEquals(0, keysToDelete.size());
+  }
+
+  @Test
+  public void bothCollisionTest() {
+
+    // State
+    String datasetKey = "508089ca-ddb4-4112-b2cb-cb1bff8f39ad";
+
+    String occId = "5760d633-2efa-4359-bbbb-635f7c200803";
+    String triplet = OccurrenceKeyBuilder.buildKey("ic", "cc", "cn").orElse(null);
+    long gbifId = 1L;
+
+    String lookupOccId = datasetKey + "|" + occId;
+    String lookupTriplet = datasetKey + "|" + triplet;
+
+    String occId2 = "5760d633-2efa-4359-bbbb-635f7c200804";
+    long gbifId2 = 2L;
+
+    HbaseStore.putRecords(
+        HBASE_SERVER.getLookupTableStore(),
+        KV.create(lookupOccId, gbifId),
+        KV.create(lookupTriplet, gbifId2),
+        KV.create(occId2, gbifId2));
+
+    HBaseLockingKeyService keygenService =
+        new HBaseLockingKeyService(HbaseServer.CFG, HBASE_SERVER.getConnection(), datasetKey);
+
+    // When
+    Map<String, Long> keysToDelete =
+        new BothStrategy().getKeysToDelete(keygenService, true, triplet, occId);
+
+    // Should
+    Assert.assertEquals(2, keysToDelete.size());
+    Assert.assertEquals(Long.valueOf(gbifId), keysToDelete.get(occId));
+    Assert.assertEquals(Long.valueOf(gbifId2), keysToDelete.get(triplet));
   }
 }
