+++ gbif/identifiers/identifiers-integration-tests/src/test/java/org/gbif/pipelines/diagnostics/tools/IdentifiersMigratorToolIT.java
@@ -4,8 +4,8 @@ import java.io.IOException;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.UUID;
-import org.gbif.pipelines.diagnostics.resources.HbaseServer;
 import org.gbif.pipelines.keygen.api.KeyLookupResult;
+import org.gbif.pipelines.resources.HbaseServer;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.ClassRule;
