+++ gbif/identifiers/identifiers-integration-tests/src/test/java/org/gbif/pipelines/diagnostics/tools/RepairGbifIDLookupToolIT.java
@@ -6,11 +6,11 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Optional;
 import org.gbif.pipelines.diagnostics.MainTool;
-import org.gbif.pipelines.diagnostics.resources.HbaseServer;
 import org.gbif.pipelines.diagnostics.strategy.DeletionStrategy.DeletionStrategyType;
 import org.gbif.pipelines.diagnostics.strategy.LookupKeyUtils;
 import org.gbif.pipelines.keygen.HBaseLockingKeyService;
 import org.gbif.pipelines.keygen.identifier.OccurrenceKeyBuilder;
+import org.gbif.pipelines.resources.HbaseServer;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.ClassRule;
