+++ gbif/identifiers/identifiers-integration-tests/src/test/resources/dwca/regular/meta.xml
