+++ gbif/identifiers/identifiers-integration-tests/src/test/resources/dwca/regular/occurrence.txt
