+++ gbif/identifiers/identifiers-integration-tests/src/test/resources/logback-test.xml
